Change fuel thesaurus format and add test.
Major change to fuel thesaurus format.
Each major fuel type is now in a stand-alone file with the standard name
(used on our database) as the first line and all subsequent aliases in the
remaining lines.
The directory for the fuel thesaurus is resources/fuel_types_thesaurus.

As part of a testing suite to assert that multiple fuel types are not
expressed under the same or different standard fuel, duplicate entries
have been removed from the fuel thesaurus files.
Note that the lowercase form of the standard fuel name is not included as an
alias in the file but is included in `make_fuel_type_thesaurus()`.

Some strange aliases remain, included "coal -  petcoke" under "Gas" and
"petroleum" under "Coal". These will need to be investigated.

Tests are located in the directory utils/tests.