Change generation system to store past 5 years.
The generation storage system is redesigned, allowing for
multiple generation values to be associated with any single plant.
This is done by storing generation data in a PlantGenerationObject,
which stores GWH, start and end dates for the GWH value,
and a source that describes where the generation data originated.

This change caused other breaking changes, which are resolved in this
commit. The output CSV now features 5 columns for each of the past
5 years, replacing the previous format of individual generation
and year_of_generation columns.

The summarizing script is also updated to reflect this change,
as well as fix a previous error caused by the recent addition
of the fuel type petcoke.