Update local databases following new generation method.
The local databses in source_databases and source_databases_csv
are updated following the change to the generation method in the
previous commit.

The output CSV database is updated, including the new generation
columns and removing the old generation columns.
Most importantly, the output database is now sorted by
country and then plant name. This should make perusing it easier!