Fix WRI database building with correct idnr
Previous change to build_database_WRI.py caused pw_idnr to not include
the WRI prefix, resulting in errors down the road and causing most of the
WRI data to not be captured in the final database.

NOTE: The build_database_* scripts still need to be ran before running
build_powerwatch.py. There are latent changes in the codebase that
have not been applied to the pickled datasets.